+++ iosApp/iosApp/MiamManager.swift
@@ -32,6 +32,8 @@ public class MiamManager {
     // need to be private
     private init() {
         KoinKt.doInitKoin()
+        ToasterHandler.shared.setOnAddRecipeText(message: <#T##String#>)
+        ToasterHandler.shared.setOnLikeRecipeText(message: <#T##String#>)
         LogHandler.companion.info("Are you ready ? \(ContextHandlerInstance.shared.instance.isReady())")
         ContextHandlerInstance.shared.instance.onReadyEvent(callback: {isReady in print("Miam event recived \(isReady)")})
         basketHandler = BasketHandlerInstance.shared.instance
@@ -102,7 +104,7 @@ public class MiamManager {
     }
     
     private func initTemplate(){
-        Template.sharedInstance.counterTemplate =
+        Template.sharedInstance.counterViewTemplate =
         {(count: Int,
           increase: @escaping () -> Void,
           decrease: @escaping () -> Void ) -> AnyView in
@@ -123,4 +125,3 @@ public class MiamManager {
             )}
     }
 }
-