+++ miamCore/src/commonMain/kotlin/com/miam/kmmMiamCore/miam_core/data/repository/RecipeRepositoryImp.kt
@@ -52,17 +52,26 @@ class RecipeRepositoryImp(
         return addRecipeLikes(recipes)
     }
 
-    override suspend fun getRecipeSuggestions(
+    override suspend fun getRecipeSuggestion(
         supplierId: Int,
         criteria: SuggestionsCriteria
     ): Recipe {
+        // TODO : WARNING what if list is empty !!??
+        return getRecipeSuggestions(supplierId, criteria, 1)[0]
+    }
+
+    override suspend fun getRecipeSuggestions(
+        supplierId: Int,
+        criteria: SuggestionsCriteria,
+        size: Int
+    ): List<Recipe> {
         val recipes = recipeDataSource.getRecipeSuggestions(
             supplierId,
+            size,
             criteria,
             listOf("ingredients", "recipe-steps", "recipe-provider", "recipe-status", "recipe-type")
         )
-        // TODO : WARNING what if list is empty !!??
-        return addRecipeLike(recipes[0])
+        return recipes.map { addRecipeLike(it) }
     }
 
     suspend fun addRecipeLikes(recipeList: List<Recipe>): List<Recipe> {