+++ changelog.md
@@ -1,6 +1,23 @@
 #### 3.0.0
 - [FEA][AND] Refact catalog header
 
+## 2.5.2
+- [FIX][IOS] Pass browse catalog action to favorites empty view template
+
+## 2.5.1
+- [FIX][IOS] Add missing browse catalog action in FavoritesView
+
+## 2.5.0
+- [FEA][COR] Prevent infinit loading on basket preview if miam is not well configure
+- [FIX][IOS] Add missing template for recipe card loading view
+- [FEA][IOS] Add a parameter to control the height of recipe card view. 
+- [FEA][IOS] Recipe carousel
+
+## 2.4.0
+- [FEA][COR] carousel view model
+- [FEA][AND] recipe carousel component
+- [FIX][IOS] Infinite loading on recipe card when using a template
+- [FIX][IOS] Add missing templates
 
 ## 2.3.0
 - [FIX][IOS] Fix Auto resize for UIkit table cells
@@ -21,6 +38,10 @@
 - [FIX][COR] Gests counter block at 1 or 100 
 - [FIX][COR] Dynamic stand alone favorite page controleur
 - [FIX][COR] Analytics add pos id to confirm event
+- [FIX][COR] add params to template
+
+## 2.2.7
+- [FIX][COR] Avoid div by zero
 
 ## 2.2.6
 - [FIX][AND] Difficulty checkbox in filter are now working